move validator into src/
